move parse_size
