+++ xarray/plot/utils.py
@@ -1224,3 +1224,44 @@ def _infer_meta_data(ds, x, y, hue, hue_style, add_guide, funcname):
         "ylabel": label_from_attrs(ds[y]),
         "hue": hue,
     }
+
+
+# copied from seaborn
+def _parse_size(data, norm):
+
+    import matplotlib as mpl
+
+    if data is None:
+        return None
+
+    data = data.values.flatten()
+
+    if not _is_numeric(data):
+        levels = np.unique(data)
+        numbers = np.arange(1, 1 + len(levels))[::-1]
+    else:
+        levels = numbers = np.sort(np.unique(data))
+
+    min_width, max_width = _MARKERSIZE_RANGE
+    # width_range = min_width, max_width
+
+    if norm is None:
+        norm = mpl.colors.Normalize()
+    elif isinstance(norm, tuple):
+        norm = mpl.colors.Normalize(*norm)
+    elif not isinstance(norm, mpl.colors.Normalize):
+        err = "``size_norm`` must be None, tuple, or Normalize object."
+        raise ValueError(err)
+
+    norm.clip = True
+    if not norm.scaled():
+        norm(np.asarray(numbers))
+    # limits = norm.vmin, norm.vmax
+
+    scl = norm(numbers)
+    widths = np.asarray(min_width + scl * (max_width - min_width))
+    if scl.mask.any():
+        widths[scl.mask] = 0
+    sizes = dict(zip(levels, widths))
+
+    return pd.Series(sizes)
